Feature/context manager

+++ scripts/create_milvus_collections.py
@@ -1,3 +1,5 @@
+import argparse
+
 from pymilvus import (
     connections,
     Collection,
@@ -99,6 +101,7 @@ def create_categories_collection():
         # Adjust dimension as needed. Use 1024 for intfloat/e5-large-v2
         # Use 768 for sentence-transformers/multi-qa-mpnet-base-dot-v1
         FieldSchema(name="embedding", dtype=DataType.FLOAT_VECTOR, dim=1024), # Adjust dimension as needed
+        FieldSchema(name="full_embedding", dtype=DataType.FLOAT_VECTOR, dim=1024), # Adjust dimension as needed
 
     ]
     
@@ -124,10 +127,22 @@ def create_categories_collection():
         index_params=index_params
     )
     
+    collection.create_index(
+        field_name="full_embedding",
+        index_params=index_params
+    )
+
     print(f"✅ Created collection: categories_collection")
     return collection
 
 def main():
+
+    parser = argparse.ArgumentParser(description="Embedding Generation Script")
+    parser.add_argument("option", type=str, help="Products, KnowledgeBase, Categories, or All")
+    args = parser.parse_args()
+    option = args.option.lower()
+    
+
    # Connect to Milvus
     connections.connect(
         alias="default",
@@ -137,23 +152,25 @@ def main():
     
     print("🔗 Connected to Milvus")
     
-    # Drop existing collections if they exist
-    if utility.has_collection("products_collection"):
-        utility.drop_collection("products_collection")
-        print("Dropped existing products_collection")
+    if option in ["products", "all"]:
+        # Drop existing collections if they exist
+        if utility.has_collection("products_collection"):
+            utility.drop_collection("products_collection")
+            print("Dropped existing products_collection")
+        create_products_collection()
     
-    if utility.has_collection("knowledge_base_collection"):
-        utility.drop_collection("knowledge_base_collection")
-        print("Dropped existing knowledge_base_collection")
-    
-    if utility.has_collection("categories_collection"):
-        utility.drop_collection("categories_collection")
-        print("Dropped existing categories_collection")
+    if option in ["knowledgebase", "all"]:
+        if utility.has_collection("knowledge_base_collection"):
+            utility.drop_collection("knowledge_base_collection")
+            print("Dropped existing knowledge_base_collection")
+        create_knowledge_base_collection()
 
-    # Create collections
-    create_products_collection()
-    create_knowledge_base_collection()
-    create_categories_collection()
+    if option in ["categories", "all"]:
+        if utility.has_collection("categories_collection"):
+            utility.drop_collection("categories_collection")
+            print("Dropped existing categories_collection")
+
+        create_categories_collection()
     
     # List collections
     collections = utility.list_collections()
